pyproject.toml
[tool.poetry]
name = "fadin"
version = "0.1.1"
description = ""
authors = ["Robert Urbańczyk <robertu@robertu.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
fastapi = "^0.104.1"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
django-extensions = "^3.2.3"
uvicorn = "^0.24.0.post1"
python-multipart = "^0.0.6"
ujson = "^5.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
whitenoise = "^6.6.0"
pytest-django = "^4.7.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
ipython = "^8.18.0"
pytest-factoryboy = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
pythonpath = "fadin"
addopts = "-p no:warnings"
testpaths = [ "tests/**/test*.py" ]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)-8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"